feat(summit-rcm): Add live REST API documentation
This change introduces a new, optional feature which enables live documentation
for the Summit RCM REST API.

As part of the build process, an OpenAPI spec document (JSON) is generated and
copied to the rootfs. When a user navigates to the API doc URL
(http://<IP ADDRESS OF THE SOM>/api/docs), Summit RCM serves up an HTML document
which references in a JavaScript library (Scalar) which is available online and
passes the OpenAPI spec to it. Scalar, which runs completely from the user's
browser, analyzes the OpenAPI spec and displays a modern and easy-to-use
interface which provides support for interacting with the Summit RCM REST API
"live" and also provides example code-snippets for a plethora of languages.

Refs: PROD-2280