fix(VSP): Fix D-Bus signal handler cleanup
The underlying dbus-fast library used to register a signal handler for the
device's PropertiesChanged signal stores the registered handlers as instance
members per-object. The current code makes the (incorrect) assumption that
subsequent calls to the 'get_interface()' function will return a single
instance of the interface, but this is not the case.

Therefore, this change now stores the device's property interface as a
member of the 'VspConnection' object and references it again during
connection cleanup to properly remove the signal handler.

Refs: PROD-7636