fix(VSP): Keep VSP GATT server alive if client disconnects
While the VSP GATT server currently only supports one client at a time, the
logic currently assumes that the server should close if an existing client
disconnects.

This change modifies the logic to allow the VSP GATT server to remain running if
the current client disconnects (to allow for subsequent client connections) and
to also close an existing connection if a new client connects (and switch to the
new one).

Refs: PROD-7588