fix(VSP): Fix detecting 'auth failure' DisconnectReason
With dbus-fast and BlueZ, the dictionary returned to the device's
PropertiesChanged signal handler contains Variant-typed data; however, the logic
to detect the 'auth failure' DisconnectReason did not account for this.

This change converts the Variant-typed data returned for DisconnectReason into a
Python string so that it can be properly detected.

Refs: PROD-7581