feat(datetime): Allow setting time via datetime string
This change introduces logic to allow a user to pass a datetime string while setting the time in lieu of a microsecond UTC timestamp as an integer. To support this, the set_time_manual() function first attempts to parse the incoming string as an integer. If successful, we assume the incoming value is a timestamp and pass it along to the systemd SetTime() D-Bus function. If the parsing fails, we attempt to parse the incoming value as a datetime string, retrieve the corresponding microsecond UTC timestamp integer from it, and pass the result on to the SetTime() function.

Refs: PROD-10737