fix(bluetooth): Cleanup GetConnInfo logic
This change adds logic to first check if the target device is connected before attempting to retrieve connection info for it. If it's not connected, an error (matching the same error used in weblcm) is returned.

Additionally, this change switches to calling the GetConnInfo() function using the dbus_fast wrapper instead of via low-level MessageBus call.

Refs: PROD-12229