fix(bt: getConnInfo command): Improve error message
This change adds explicit handling for the case where there device is not connected and not paired in the getConnInfo command handler. In this case, we now return the "Device not found" message; whereas, in the past, this would throw an exception which is subsequently caught by the generic exception handler resulting in the generic "Unable to get connection info" message being returned.

Refs: PROD-13265