feat(APs): Support querying last scan time
This change includes the missing "LastScan" property for a Wireless device as
reported by NetworkManager. This property is described as "[t]he timestamp (in
CLOCK_BOOTTIME milliseconds) for the last finished network scan. A value of -1
means the device never scanned for access points."

https://networkmanager.dev/docs/api/latest/gdbus-org.freedesktop.NetworkManager.Device.Wireless.html#gdbus-property-org-freedesktop-NetworkManager-Device-Wireless.LastScan

Additionally, in order to more-easily interpret what this property practically
means for a user of the Summit RCM REST API, this change introduces a new
endpoint:

GET to /api/v2/network/accessPoints/scan

This endpoint interprets the "LastScan" property and returns the seconds since
the last scan.

For legacy endpoints, this is also now returned as an additional entry in the
response data for the /accesspoints endpoint.

Refs: PROD-16249