feat: Support configuring webserver log level
This change adds support for modifying the log level utilized by Uvicorn (the
webserver used by Summit RCM) on-the-fly via a set of new REST API endpoints.

Legacy:
- GET /logWebserver
- PUT /logWebserver

v2:
- GET /api/v2/system/logs/webserver
- PUT /api/v2/system/logs/webserver

The initial, default log level value is unchanged (error); however, this can be
configured in the /etc/summit-rcm/summit-rcm-settings.ini file on-device.

When a new log level is requested, it is persistently updated (in the
/etc/summit-rcm/summit-rcm-settings.ini file), and the new value is passed to
the running uvicorn server instance (which updates the log level on-the-fly).

The change also includes the necessary modifications to support exposing the new
REST endpoints in the live REST API documentation. Also added are a set of
example shell scripts to demonstrate exercising the endpoints.

Currently, the supported log levels (as defined by uvicorn) are:
- critical
- error (Summit RCM default)
- warning
- info
- debug
- trace

https://www.uvicorn.org/settings/#logging

Refs: PROD-15972