feat(REST: network: interface): Add available AP channels endpoint
This change introduces new REST endpoints (legacy and v2) to support returning
the subset of channels available for use in AP mode.

- legacy endpoint: `/networkInterfaceAvailableApChannels`
- v2 endpoint: `/api/v2/network/interfaces/{name}/availableApChannels`

The criteria for an "available" channel is:
- Marked as "supported" (i.e., like in `iw phy0 info`)
- Not marked as "disabled" (i.e., like in `iw phy0 info`)
- Not marked as "DFS" (i.e., like in `iw reg get`)
- Not marked as requiring passive-scan (i.e., like in `iw reg get`)

Also included are updates to the live REST API documentation to describe the
response parameters.

Additionally, a set of usage example shell scripts are added as well.

Refs: PROD-6309