fix(network service: freq info): Fix retrieving freq info
This change fixes the logic to retrieve the current operating frequency for a
given interface. Previously, the frequency of the first entry in the scan list
was used, but this is now replaced with the actual frequency for the interface
as reported by Netlink.

Additionally, this change adds a new "sibling" parameter (`channel`) which
decodes the frequency into the corresponding channel number.

Refs: PROD-19110