feat(network: interfaces): Add DHCP leases endpoint
This change introduces new REST endpoints (legacy and v2) to support returning
the set of DHCP leases which have been handed out by dnsmasq (the DHCP server
used in conjunction with NetworkManager when the 'ipv4.method' setting is set
to 'shared' on a connection profile).

- legacy endpoint: `/networkInterfaceDhcpLeases`
- v2 endpoint: `/api/v2/network/interfaces/{name}/dhcpLeases`

The DHCP leases are read from the per-interface leasefile located at
`/var/lib/NetworkManager/dnsmasq-{name}.leases` which is maintained by dnsmasq.

Also included are updates to the live REST API documentation to describe the
response parameters.

Additionally, a set of usage example shell scripts are added as well.

Refs: PROD-6308