fix(network_manager_service.py): Handle missing prop interface
This change explicitly handles the `InterfaceNotFoundError` when attempting to
retrieve the standard DBus properties interface for an object in the
`subscribe()` function.

This error can occur when the object at the given object path is "obsolete"
(i.e., does not implement the standard DBus properties interface) by the time
Summit RCM attempts to access it. One example of this is the `ActiveAccessPoint`
property of a "Wireless" device which changes rapidly when bringing up an AP
connection.

Refs: PROD-33702